Made auto_optimizers folder and move optimization code to it
