+++ forecasting_tools/__init__.py
@@ -53,6 +53,9 @@ from forecasting_tools.ai_models.general_llm import GeneralLlm as GeneralLlm
 from forecasting_tools.ai_models.resource_managers.monetary_cost_manager import (
     MonetaryCostManager as MonetaryCostManager,
 )
+from forecasting_tools.auto_optimizers.bot_optimizer import (
+    BotOptimizer as BotOptimizer,
+)
 from forecasting_tools.benchmarking.benchmark_displayer import (
     run_benchmark_streamlit_page as run_benchmark_streamlit_page,
 )
@@ -61,9 +64,6 @@ from forecasting_tools.benchmarking.benchmark_for_bot import (
 )
 from forecasting_tools.benchmarking.benchmarker import (
     Benchmarker as Benchmarker,
-)
-from forecasting_tools.benchmarking.bot_optimizer import (
-    BotOptimizer as BotOptimizer,
 )
 from forecasting_tools.data_models.binary_report import (
     BinaryReport as BinaryReport,
